+++ lib/capybara/session.rb
@@ -84,7 +84,9 @@ module Capybara
         yield config
       end
       @server = if config.run_server && @app && driver.needs_server?
-        Capybara::Server.new(@app, port: config.server_port, host: config.server_host, reportable_errors: config.server_errors).boot
+        server_options = { port: config.server_port, host: config.server_host, reportable_errors: config.server_errors }
+        server_options[:extra_middleware] = [Capybara::Server::AnimationDisabler] if config.disable_animation
+        Capybara::Server.new(@app, server_options).boot
       end
       @touched = false
     end